緊急聯絡修正
修正閃退問題
拍照給教官功能

+++ NTOU_Mobile/module/ShuttleTrack/Bus/TPRouteDetailViewController.m
@@ -29,7 +29,7 @@
 - (void) setter_busName:(NSString *)name andGoBack:(NSInteger)goback
 {
     busName = name;
-    goBack = [[NSString alloc] initWithFormat:@"%lu", goback];
+    goBack = [[NSString alloc] initWithFormat:@"%lu", (long int)goback];
     NSLog(@"busName:%@, goBack:%@", busName, goBack);
     ISREAL = FALSE;
 }
